python-services/data-analyzer-service/requirements.txt
# Web Framework & Core
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
loguru>=0.7.2
python-dotenv>=1.0.0

# LLM & AI Core
openai>=1.60.0

# Data Science Utils
numpy
pandas
scikit-learn

# PII Protection
presidio-analyzer
presidio-anonymizer
faker

# Embedding & Clustering
sentence-transformers
hdbscan
umap-learn

# Vector Database
faiss-cpu
chromadb

# LangChain Framework
langchain
langchain-openai
langgraph

# Additional Tools
tiktoken
tenacity